backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Supabase and PostgreSQL dependencies
supabase>=2.0.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
asyncpg>=0.29.0

# Database ORM and spatial data
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0
alembic>=1.12.0

# Additional utilities
requests>=2.31.0
python-jose[cryptography]>=3.3.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: for enhanced JSON handling
orjson>=3.9.0
